% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_diag.R
\name{brms_diag}
\alias{brms_diag}
\title{Create diagnostic summaries for `brms` models}
\usage{
brms_diag(fit, max_treedepth = 10)
}
\arguments{
\item{fit}{An object of class `brmsfit`.}

\item{max_treedepth}{A single integerish value indicating the first value
that is reported as "reaching the maximum treedepth". Default is 10. It is
included as a parameter because I could not find a reliable way to extract
the value from the `fit` object.}
}
\value{
A list with six elements named `div`, `tree`, `bfmi`, `rhat`,
`ess_bulk`, and `ess_tail` containing one integer each. See details for more
information.
}
\description{
Create diagnostic summaries for `brms` models
}
\details{
\describe{
\item{Divergencies (div)}{Number of divergencies}
\item{Tree depth (tree)}{Number of interations that reached the maximum tree
depth}
\item{E-BFMI (bfmi)}{Number of chains that have a value below 0.2. The
maximum number is the number of chains, because E-BFMI is evaluated by chain,
not by iteration.}
\item{Rhat (rhat)}{Number of parameters with a Rhat value above 1.01}
\item{Bulk effective sample size (ess_bulk)}{Number of parameters with a bulk
effective sample size below 100 multiplied by the number of chains, i.e.,
below 400 for if 4 chains have been used}
\item{Tail effective sample size (ess_tail)}{Number of parameters with a tail
effective sample size below 100 multiplied by the number of chains, i.e.,
below 400 for if 4 chains have been used}
}
}
\examples{
d <- data.frame(y = c(0, 1, 0, 1, 0, 1, 1, 0, 1, 1))
f <- brms::brm(
  y ~ 1, data = d, family = brms::bernoulli(),
  refresh = 0, silent = 2
)
brms_diag(f)
}
