% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{create_hmm_data}
\alias{create_hmm_data}
\title{Create simple data for testing}
\usage{
create_hmm_data(
  use_state = TRUE,
  use_time = TRUE,
  to_factor = TRUE,
  na_obs = TRUE,
  na_cov = TRUE
)
}
\arguments{
\item{use_state}{TRUE (default) or FALSE, indicating if a `state` variable is
supposed to be included for semi-supervised HMMs}

\item{use_time}{TRUE (default) or FALSE, indicating if a `time` variable is
supposed to be included}

\item{to_factor}{TRUE (default) or FALSE, indicating if the categorical
covariates should be transormed to factor}

\item{na_obs}{TRUE (default) or FALSE, indicating if the variables caused by
the latent states include NAs}

\item{na_cov}{TRUE (default) or FALSE, indicating if the covariats include
NAs}
}
\value{
A data.frame that can be used in `run_hmm()`
}
\description{
Create simple data for testing
}
\examples{
df <- create_hmm_data()
str(df)
}
