% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{extract.symbols.from.ast}
\alias{extract.symbols.from.ast}
\title{Extract symbols from AST}
\usage{
extract.symbols.from.ast(x)
}
\arguments{
\item{x}{An expression (formed, e.g., by wrapping inside of `rlang::expr()`)}
}
\value{
A character vector containing all symbol names. If no symobl are in
the call, returns an empty character vector, i.e., character(0).
}
\description{
Extract all symbols from abstract syntax trees.
}
\examples{
extract.symbols.from.ast(rlang::expr(is.na(y)))
}
