% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{get.all.functions}
\alias{get.all.functions}
\title{Get all function and package names}
\usage{
get.all.functions(include.package.names = TRUE)
}
\arguments{
\item{include.package.names}{TRUE (default) or FALSE, indicating if names of
packages are supposed to be included as well.}
}
\value{
A (usually large) character vector containing the names of all
available functions and (potentially) packages.
}
\description{
Get all function and package names
}
\examples{
get.all.functions()
}
