% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lca_example_data}
\alias{lca_example_data}
\title{LCA example data}
\format{
## `lca_example_data`
A data.frame with 500 rows and 6 columns
\describe{
 \item{var1}{First indicator variable}
 \item{var2}{Second indicator variable}
 \item{var3}{Third indicator variable}
 \item{var4}{Fourth indicator variable}
 \item{sex}{A categorical covariate}
 \item{bmi}{A continuous covariate}
}
}
\source{
Simulated data
}
\usage{
lca_example_data
}
\description{
A dataset with simulated values for testing and examples
}
\keyword{datasets}
