% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca_ident}
\alias{lca_ident}
\title{Check LCA model identification}
\usage{
lca_ident(lca, for_table = F)
}
\arguments{
\item{lca}{An object of class `poLCA` created by `poLCA::poLCA()`}

\item{for_table}{TRUE or FALSE indicating if the result should be used for
`supfuns::lca_extract()`}
}
\value{
A single value indicating the percentage of starting values
resulting in the best model
}
\description{
Check LCA model identification
}
\examples{
f <- cbind(var1, var2, var3, var4) ~ 1
lca <- poLCA::poLCA(f, lca_example_data, nclass = 3, verbose = FALSE)
lca_ident(lca)
}
