% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca_plot_input}
\alias{lca_plot_input}
\title{Extracts the information necessary for plotting the lca}
\usage{
lca_plot_input(lca)
}
\arguments{
\item{lca}{An object of class `poLCA` created by `poLCA::poLCA()`}
}
\value{
A table that can be used for creating a plot that describes the LCA
}
\description{
Extracts the information necessary for plotting the lca
}
\examples{
f <- cbind(var1, var2, var3, var4) ~ 1
lca <- poLCA::poLCA(f, lca_example_data, nclass = 3, verbose = FALSE)
lca_plot_input(lca)
}
