% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca_print}
\alias{lca_print}
\title{Print an object of class `poLCA`}
\usage{
lca_print(lca, cut = FALSE, nice = TRUE)
}
\arguments{
\item{lca}{An object of class `poLCA` created by `poLCA::poLCA()`}

\item{cut}{TRUE or FALSE indicating if values below 0 or above 1 in the
confidence interval (CI) should be replaced by 0 and 1}

\item{nice}{TRUE or FALSE indicating if estimate and CI should be formatted
nicely}
}
\value{
Print a nice summary of the object passed to `lca`
}
\description{
Print an object of class `poLCA`
}
\examples{
f <- cbind(var1, var2, var3, var4) ~ 1
lca <- poLCA::poLCA(f, lca_example_data, nclass = 3, verbose = FALSE)
lca_print(lca)
}
