% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca_sample}
\alias{lca_sample}
\title{Sample class assignment}
\usage{
lca_sample(lca, ndraws = 20)
}
\arguments{
\item{lca}{An object of class `poLCA` created by `poLCA::poLCA()`}

\item{ndraws}{The number of random draws of latent class assignment}
}
\value{
A data.frame with `ndraws` variables representing one random draw of
latent class assignments each. Variables are factors with labels `lc1`,
`lc2`, etc.
}
\description{
Sample class assignment
}
\examples{
f <- cbind(var1, var2, var3, var4) ~ 1
lca <- poLCA::poLCA(f, lca_example_data, nclass = 3, verbose = FALSE)
df <- cbind(lca_example_data, lca_sample(lca))
}
