% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca_table}
\alias{lca_table}
\title{Table creation latent class model}
\usage{
lca_table(lca_list)
}
\arguments{
\item{lca_list}{A list of objects of class `poLCA` created by
`poLCA::poLCA()`}
}
\value{
A table containing information on the models
}
\description{
Table creation latent class model
}
\examples{
f <- cbind(var1, var2, var3, var4) ~ 1
lca2 <- poLCA::poLCA(f, lca_example_data, nclass = 2, nrep = 100, verbose = FALSE)
lca3 <- poLCA::poLCA(f, lca_example_data, nclass = 3, nrep = 100, verbose = FALSE)
lca_table(list(lca2, lca3))
}
