% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi_diag_cat.R
\name{mi_diag_cat}
\alias{mi_diag_cat}
\title{Diagnose imputation of categorical variables}
\usage{
mi_diag_cat(imp, var)
}
\arguments{
\item{imp}{A `mids` object, i.e., the output of function `mice::mice()`}

\item{var}{A single string containing the name of the variable to be plotted.
`var` must be factor.}
}
\value{
A ggplot object
}
\description{
Compare the proportions of categories in non-imputed and imputed datasets in
a plot
}
\examples{
df <- data.frame(
x = rnorm(100, 0, 1),
y = sample(c(1, 2, 3), size = 100, replace = TRUE, prob = c(0.2, 0.3, 0.5)),
r = runif(100, 0, 1)
) \%>\%
  dplyr::mutate(
    "y" := dplyr::case_when(
      .data$r < 0.05 ~ NA,
      .default = .data$y
    ),
    "r" := NULL,
    "y" := factor(.data$y)
  )
imp <- mice::mice(df, m = 2, maxit = 2)
mi_diag_cat(imp, "y")
}
