% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{n_trans}
\alias{n_trans}
\title{Derive number of transitions}
\usage{
n_trans(hmm, nsamp = 1, diag_mode = FALSE)
}
\arguments{
\item{hmm}{A `hmmTMB::HMM` object, which has already been fitted with
`HMM$fit()`}

\item{nsamp}{An integerish number indicating the number of state sequences
to be generated. If `nsamp` == 1, HMM$viterbi() is used. Otherwise,
HMM$sample_states() is used.}

\item{diag_mode}{TRUE or FALSE (default) indicating if diagnostic mode is
activated}
}
\value{
A list with elements `rem`, `rel`, and `by_id`. `rem` and `rel` are
numeric vectors with length equal to `nsamp` and contain the number of
remission and relapse events, respectively. `by_id` is a data.frame with
`nsamp + 1` columns. The first column is called `ID` and contains participant
IDs, which are the unique values of variable `ID` in the data passed to the
`hmmTMB` objects. There is one row for each unique value in `ID`. The
remaining columns have name `remrel<number>` with `<number>` being replaced
by every value from 1 to `nsamp`, e.g., `remrel1` for the first of them.
They are of type character and contain a value that is a combination of two
digits: the first one represents the number of remission events, while the
second represents the number of relapse events. Remember the order of
remission first, relapse second with the variable name `remrel`. For example,
the value `21` means that the individual had 2 remission events and 1
relapse event.
}
\description{
Derive number of transitions
}
\examples{
df <- create_hmm_data()
x <- run_hmm(df, obsvars = c("obs1", "obs2"), nrep = 1)
n_trans(x$hmm)
}
