% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{or_hmm}
\alias{or_hmm}
\title{Extract estimates for transition probability parameters}
\usage{
or_hmm(hmm, var = NULL)
}
\arguments{
\item{hmm}{A `hmmTMB::HMM` object, which has already been fitted with
`HMM$fit()`}

\item{var}{A character vector containing the list of variables of interest.
All parameters that contain at least one of the strings in `var` are included
in the output table.}
}
\value{
A data.frame with estimates for transition parameters
}
\description{
Extract estimates for transition probability parameters
}
\examples{
df <- create_hmm_data()
x <- run_hmm(df, obsvars = c("obs1", "obs2"), nrep = 2, formula = ~female)
or_hmm(x$hmm, var = "female")
}
