% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{plot_hmm_obs}
\alias{plot_hmm_obs}
\title{Plot probabilities of observed variables by state}
\usage{
plot_hmm_obs(hmm)
}
\arguments{
\item{hmm}{A `hmmTMB::HMM` object, which has already been fitted with
`HMM$fit()`}
}
\value{
A ggplot object
}
\description{
Creates a plot with one pointrange for each variable and state. In the
project context, states are "Asthma" and "No Asthma"
}
\examples{
df <- create_hmm_data()
x <- run_hmm(df, obsvars = c("obs1", "obs2"), nrep = 2)
plot_hmm_obs(x$hmm)
}
