% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{plot_hmm_predict}
\alias{plot_hmm_predict}
\title{Plot predictions of transitions probabilities}
\usage{
plot_hmm_predict(
  hmm,
  var = NULL,
  facet = NULL,
  fix = NULL,
  n_post = 1000,
  diag_mode = FALSE
)
}
\arguments{
\item{hmm}{A `hmmTMB::HMM` object, which has already been fitted with
`HMM$fit()`}

\item{var}{A character string of length 1 containing the name of the variable
for which a comparison is shown inside the same plot}

\item{facet}{A formula passed to `facet_wrap()` or `facet_grid()`}

\item{fix}{A named list of length 1 vectors indicating which variables
(names of list elements) should be fixed to which values (vectors in list
elements) when creating the new data to pass to `HMM$predict()`, based on
which the plots are created. Variables that are not in `var` or `facet` need
to be fixed.}

\item{n_post}{A single integerish value indicating how many posterior samples
are drawn by `HMM$predict()`}

\item{diag_mode}{TRUE or FALSE (default) indicating if diagnostic mode is
activated}
}
\value{
A list of two ggplot objects, `rem` and `rel`, holding the plots for
the effect on the probability of remission and relapse, respectively. If
`diag_mode = TRUE`, the list additionally contains elements `newdata`,
containing the element of the same name passed to `hmmTMB::HMM$predict()`,
`age_sex`, containing the variable names of the age and sex variables, `var`,
containing the name of the variable, for which the comparison of its
categories is included within each subplot, `facet_vars`, containing the
variable names used for creating subplots, and `fix_vars`, including the
names of variables, for which only a single value is included in `newdata`.
}
\description{
Plot predictions of transitions probabilities
}
\examples{
df <- create_hmm_data()
x <- run_hmm(df,
  obsvars = c("obs1", "obs2"), nrep = 1,
  formula = ~ age_pred * female + vab1
)
plot_hmm_predict(x$hmm, var = "vab1", facet = ~female, n_post = 1e2)
}
