% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{rdiscunif}
\alias{rdiscunif}
\title{Sample from a discrete-uniform distribution}
\usage{
rdiscunif(n, min, max)
}
\arguments{
\item{n}{A single integerish value representing the number of samples}

\item{min}{A vector of integerish values, either of length 1 or of length
`n`, representing the lower bound of the distribution.}

\item{max}{A vector of integerish values, either of length 1 or of length
`n`, representing the upper bound of the distribution.}
}
\value{
A vector of type double of length `n` with integerish values between
`min` and `max` and equal probability for all values.
}
\description{
Create random values from a range of integers with equal probability for each
integer value in between a lower and an upper bound. The boundaries are part
of the set of possible values, too.
}
\examples{
rdiscunif(n = 1e2, min = 3, max = 12)

rdiscunif(n = 3, min = c(1, 2, 3), max = c(10, 11, 12))
}
