% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{run_hmm}
\alias{run_hmm}
\title{Create and fit a Hidden Markov Model}
\usage{
run_hmm(
  data,
  obsvars,
  initial_state = c("shared", "stationary"),
  formula = NULL,
  nrep = 100,
  diag_mode = FALSE
)
}
\arguments{
\item{data}{A data.frame which is passed to argument `data` of the
`hmmTMB::Observation` object}

\item{obsvars}{A character vector including the names of the variables
included in the observation model}

\item{initial_state}{A character string, either "shared" (default) or
"stationary" containing the value that is passed to argument `initial_state`
of the `hmmTMB::MarkovChain` object}

\item{formula}{A formula object describing the linear model of the transition
probabilities, which is passed to argument `formula` of the
`hmmTMB::MarkovChain` object. Default is `NULL` corresponding to the
intercept model without any predictors.}

\item{nrep}{A single integerish value indicating the number of repetitions
of model fitting with different starting values. Default (and recommended
value) is 100.}

\item{diag_mode}{TRUE or FALSE (default) indicating if diagnostic mode is
activated}
}
\value{
With `diag_mode = FALSE`, a list of length 2 with elements `hmm`
containing the best of the `nrep` models, and `ident`, a number between 0 and
1, indicating what percentage of the `nrep` models has the same likelihood as
the best model. `ident` of 1 indicates perfect identification because all
models led to the same maximum likelihood solution, indicating that the
global maximum has been found. In latent class analysis, we are reasonable
sure to have found the global maximum likelihood solution when 50-60%
(`ident` equals 0.5-0.6) of models, i.e., starting values, lead to the same
likelihood as the best model. With `diag_mode = TRUE`, the list additionally
contains the following elements: `diag_all_models`, a list of all models
before selecting the best one, `diag_hid_draws`, a list of the drawn starting
values for the transition parameters, `diag_hid_updated`, a list of the `HMM`
object's values for the transition parameters after updating the starting
values, `diag_obs_draws`, a list of the drawn starting values for symptom
probabilities by state, `diag_obs_updated`, a list of the `HMM` object's
values for the symptom probabilities by state after updating the starting
values, `diag_unfitted_hid`, the default starting values for the transition
parameters, and `diag_unfitted_obs`, the default starting values for the
symptom probabilities by state.
}
\description{
Creates a Hidden Markov Model with `hmmTMB` and fits it with different sets
of random starting values, reporting identification and the best model.
}
\examples{
df <- create_hmm_data()
hmm <- run_hmm(df, obsvars = c("obs1", "obs2"), nrep = 1)
}
