% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.R
\name{run_sim}
\alias{run_sim}
\title{Run simluations based on simcausal modules}
\usage{
run_sim(dags, args = NULL, n = 1000, seed = NULL)
}
\arguments{
\item{dags}{A list of `simcausal::DAG` objects. The DAGs are not locked yet,
because locking is done as part of `run_sim()`. The DAGs must not include the
same variables multiple times, e.g., if a variable from module1 is needed in
module2, module 2 must not contain the variable, because when combining both
modules in this list, the variable would be available twice leading to an
error. Similarly, the order, in which DAGs are listed is important, e.g.,
module1 must be listed first because its variable that is used in module2
would otherwise not be available.}

\item{args}{A named list of all parameters used in `dags`, which are not
other variables in the DAGs. If, e.g., a variable is described be a normal
distribution with a fixed mean and this mean is called `mean_eg` in the code
that defines the DAG, then `mean_eg` must be provided in this list.}

\item{n}{An integerish number indicating the number of samples. Default is
1000.}

\item{seed}{An integerish number passed to argument `rndseed` in function
`simcausal::sim()`.}
}
\value{
A data.frame with columns `ID` as well as one column for each
variable included in the DAG. The data.frame has `n` rows. It is in wide
format, because variables that exist at different time points are different
columns in the data.frame, e.g., `asthma_6` and `asthma_7` for asthma at age
6 and 7, respectively.
}
\description{
Run simluations based on simcausal modules
}
\examples{
simtest <- simcausal::DAG.empty() \%>\%
  simcausal::add.nodes(
    simcausal::node("x",
      distr = "rnorm",
      mean = 0,
      sd = 14
    )
  )
run_sim(dags = list(simtest), n = 5)
}
