% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{seed.pls}
\alias{seed.pls}
\title{Sample seeds}
\usage{
seed.pls(n = 1, print = TRUE)
}
\arguments{
\item{n}{An integer or double representation of an integer, describing the
number of seeds to sample. Default is 1.}

\item{print}{TRUE or FALSE, describing if the seeds should be printed in a
copy-friendly format in the console or not. Default is TRUE, i.e., copy-
friendly. In order to create a vector of seeds to be used, set `print` to
FALSE.}
}
\value{
Either prints seeds in copy-friendly format in the console (`print`
= TRUE), or returns a vector of integers.
}
\description{
A simple helper function for drawing random integers for use in `set.seed()`
or `seed` arguments of functions. Let's not always use `123`!
}
\examples{
seed.pls(5)
}
